:recycle: Properly name UseCases on ProcessInstanceStateHandlingFacade
