+++ src/runtime/execute_process_service.ts
@@ -314,7 +314,7 @@ export class ExecuteProcessService implements IExecuteProcessService {
         throw error;
       });
 
-      await this.processInstanceStateHandlingFacade.saveCorrelation(identity, processInstanceConfig);
+      await this.processInstanceStateHandlingFacade.saveProcessInstance(identity, processInstanceConfig);
 
       const startEventHandler = await this.flowNodeHandlerFactory.create(processInstanceConfig.startEvent);
 
@@ -329,9 +329,9 @@ export class ExecuteProcessService implements IExecuteProcessService {
       const allResults = processInstanceConfig.processTokenFacade.getAllResults();
       const resultToken = allResults.pop();
 
-      await this.processInstanceStateHandlingFacade.finishProcessInstanceInCorrelation(identity, processInstanceConfig, resultToken);
+      await this.processInstanceStateHandlingFacade.finishProcessInstance(identity, processInstanceConfig, resultToken);
     } catch (error) {
-      await this.processInstanceStateHandlingFacade.finishProcessInstanceInCorrelationWithError(identity, processInstanceConfig, error);
+      await this.processInstanceStateHandlingFacade.finishProcessInstanceWithError(identity, processInstanceConfig, error);
 
       throw error;
     }
