+++ src/runtime/facades/process_instance_state_handling_facade.ts
@@ -43,7 +43,7 @@ export class ProcessInstanceStateHandlingFacade {
    * @param   identity              The identity of the requesting user.
    * @param   processInstanceConfig The configs for the ProcessInstance.
    */
-  public async saveCorrelation(identity: IIdentity, processInstanceConfig: IProcessInstanceConfig): Promise<void> {
+  public async saveProcessInstance(identity: IIdentity, processInstanceConfig: IProcessInstanceConfig): Promise<void> {
 
     const processDefinition = await this.processModelUseCases.getProcessDefinitionAsXmlByName(identity, processInstanceConfig.processModelId);
 
@@ -67,7 +67,7 @@ export class ProcessInstanceStateHandlingFacade {
    * @param   processInstanceConfig The configs for the ProcessInstance.
    * @param   resultToken           The result with which to finish the ProcessInstance.
    */
-  public async finishProcessInstanceInCorrelation(
+  public async finishProcessInstance(
     identity: IIdentity,
     processInstanceConfig: IProcessInstanceConfig,
     resultToken: IFlowNodeInstanceResult,
@@ -92,7 +92,7 @@ export class ProcessInstanceStateHandlingFacade {
    * @param   processInstanceConfig The configs for the ProcessInstance.
    * @param   error                 The error that occured.
    */
-  public async finishProcessInstanceInCorrelationWithError(
+  public async finishProcessInstanceWithError(
     identity: IIdentity,
     processInstanceConfig: IProcessInstanceConfig,
     error: BaseError,
